Time: 18 ms (19.93%) | Memory: 14 MB (88.82%) - LeetSync

+++ 2227-sum-of-subarray-ranges/sum-of-subarray-ranges.cpp
@@ -8,7 +8,7 @@ public:
             for(int j = i+1;j<n;j++){
                 largest = max(largest,nums[j]);
                 smallest = min(smallest , nums[j]);
-                sum  += (long long)(largest - smallest);
+                sum  += (largest - smallest);
             }
         }
         return sum;
